requirements.txt
flask==2.3.2
twilio==8.10.0
python-dotenv==1.0.0
gunicorn==21.2.0
